% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithfunc.R
\name{mul}
\alias{mul}
\title{Multiply together two numbers.}
\usage{
mul(a, b)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Multiply together two numbers.
}
\examples{
mul(1, 1)
mul(10, 1)
}
